requirements.txt
certifi==2017.4.17
chardet==3.0.4
idna==2.5
pymongo==3.4.0
PyYAML==3.12
redis==2.10.5
requests==2.18.1
urllib3==1.21.1
